pyproject.toml
[tool.poetry]
name = "go"
version = "0.1.0"
description = ""
authors = ["Carlos Huaccho <Carlos@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "*"
pyyaml = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
